requirements.txt
aiohttp==1.0.5

discord.py==0.16.7
lxml==3.7.3
Pybooru==4.1.0
pyowm==2.6.1
pyparsing==2.1.10
requests==2.13.0
requests-oauthlib==0.8.0
tweepy==3.5.0
xmltodict==0.10.2
